pyproject.toml
[build-system]
requires = ["flit_core >=3.3,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "heavyai"
version = "0.30"
requires-python = ">=3.7,<3.10"
authors = [{name = "Heavy.AI", email = "community@heavy.ai"}]
description = "Data science toolkit support for HeavyDB"
readme = "README.rst"
license = {file = "LICENSE.md"}
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Topic :: Database",
    "Topic :: Scientific/Engineering",
]

dependencies = [
    "pyomniscidb >=5.5.0",
    "pyarrow >=3.0.0",
    "pandas",
    "numpy",
    "shapely",
    "importlib_metadata",  # drop for python >3.7
]

[project.optional-dependencies]
gpu = [
    "numba >=0.49"
]

doc = [
    "sphinx",
    "numpydoc",
    "sphinx-rtd-theme"
]

test = [
    "pytest",
    "coverage",
    "pytest-mock",
    "geopandas"
]

dev = [
    "heavyai[doc]",
    "heavyai[test]",
    "pre-commit",
    "flit"
]

[project.urls]
Home = "https://github.com/heavyai/heavyai"
Source = "https://github.com/heavyai/heavyai"

[tool.flit.sdist]
exclude = [
    "tests/*", "scripts/*", "ci/*", ".github/*",
    "*.yml", "*.yaml", ".*", "Jenkinsfile", "Makefile"]

[tool.black]
line-length = 79
skip-string-normalization = true
target-version = ["py37", "py38", "py39"]
exclude = "tests/data|docs|scripts"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-rsx -v"
testpaths = [
    "tests",
]
